Debt: inline module ids instead of having some of them spread out
