+++ src/buildfile.js
@@ -3,7 +3,30 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-const { createModuleDescription, createEditorWorkerModuleDescription } = require('./vs/base/buildfile');
+/**
+ * @param {string} name
+ * @param {string[]} exclude
+ */
+function createModuleDescription(name, exclude) {
+
+	let excludes = ['vs/css', 'vs/nls'];
+	if (Array.isArray(exclude) && exclude.length > 0) {
+		excludes = excludes.concat(exclude);
+	}
+
+	return {
+		name: name,
+		include: [],
+		exclude: excludes
+	};
+}
+
+/**
+ * @param {string} name
+ */
+function createEditorWorkerModuleDescription(name) {
+	return createModuleDescription(name, ['vs/base/common/worker/simpleWorker', 'vs/editor/common/services/editorSimpleWorker']);
+}
 
 exports.base = [
 	{
@@ -28,8 +51,18 @@ exports.workerSharedProcess = [createEditorWorkerModuleDescription('vs/platform/
 exports.workerLanguageDetection = [createEditorWorkerModuleDescription('vs/workbench/services/languageDetection/browser/languageDetectionSimpleWorker')];
 exports.workerLocalFileSearch = [createEditorWorkerModuleDescription('vs/workbench/services/search/worker/localFileSearch')];
 
-exports.workbenchDesktop = require('./vs/workbench/buildfile.desktop').collectModules();
-exports.workbenchWeb = require('./vs/workbench/buildfile.web').collectModules();
+exports.workbenchDesktop = [
+	createEditorWorkerModuleDescription('vs/workbench/contrib/output/common/outputLinkComputer'),
+	createModuleDescription('vs/workbench/contrib/debug/node/telemetryApp'),
+	createModuleDescription('vs/platform/files/node/watcher/watcherMain'),
+	createModuleDescription('vs/platform/terminal/node/ptyHostMain'),
+	createModuleDescription('vs/workbench/api/node/extensionHostProcess')
+];
+
+exports.workbenchWeb = [
+	createEditorWorkerModuleDescription('vs/workbench/contrib/output/common/outputLinkComputer'),
+	createModuleDescription('vs/code/browser/workbench/workbench', ['vs/workbench/workbench.web.main'])
+];
 
 exports.keyboardMaps = [
 	createModuleDescription('vs/workbench/services/keybinding/browser/keyboardLayouts/layout.contribution.linux'),
@@ -37,6 +70,14 @@ exports.keyboardMaps = [
 	createModuleDescription('vs/workbench/services/keybinding/browser/keyboardLayouts/layout.contribution.win')
 ];
 
-exports.code = require('./vs/code/buildfile').collectModules();
+exports.code = [
+	createModuleDescription('vs/code/electron-main/main'),
+	createModuleDescription('vs/code/node/cli'),
+	createModuleDescription('vs/code/node/cliProcessMain', ['vs/code/node/cli']),
+	createModuleDescription('vs/code/electron-sandbox/issue/issueReporterMain'),
+	createModuleDescription('vs/code/electron-browser/sharedProcess/sharedProcessMain'),
+	createModuleDescription('vs/platform/driver/node/driver'),
+	createModuleDescription('vs/code/electron-sandbox/processExplorer/processExplorerMain')
+];
 
 exports.entrypoint = createModuleDescription;
