--- src/vs/base/buildfile.js
@@ -1,33 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-'use strict';
-
-/**
- * @param {string} name
- * @param {string[]} exclude
- */
-function createModuleDescription(name, exclude) {
-
-	let excludes = ['vs/css', 'vs/nls'];
-	if (Array.isArray(exclude) && exclude.length > 0) {
-		excludes = excludes.concat(exclude);
-	}
-
-	return {
-		name: name,
-		include: [],
-		exclude: excludes
-	};
-}
-
-/**
- * @param {string} name
- */
-function createEditorWorkerModuleDescription(name) {
-	return createModuleDescription(name, ['vs/base/common/worker/simpleWorker', 'vs/editor/common/services/editorSimpleWorker']);
-}
-
-exports.createModuleDescription = createModuleDescription;
-exports.createEditorWorkerModuleDescription = createEditorWorkerModuleDescription;
